Again, complete refactor to use packages
