+++ grokbot/cogs/admin_commands.py
@@ -2,7 +2,7 @@ from discord import app_commands
 from discord.ext import commands
 import discord
 import asyncio
-from grokbot.utils import tail, split_log_lines
+from goim01.grokbot.grokbot.utils import tail, split_log_lines
 from grokbot.config import BOT_OWNER_ID
 
 class AdminCommands(commands.Cog):
